Fragment with detailed bicycle without implementation of repository

+++ app/src/main/java/ru/nsu/fit/data/model/Bicycle.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "bicycles")
 data class Bicycle(
-    @PrimaryKey(autoGenerate = true) val bikeId: Long,
+    @PrimaryKey(autoGenerate = true) val bikeId: Int,
     val name: String,
     val purchasePrice: Int,
     val sellingPrice: Int?,
